requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dotenv==1.0.1
readability-lxml==0.8.1
requests==2.32.3
requests-oauthlib==2.0.0
Pillow==10.4.0
Flask==3.0.3
